debian: update the efreet packaging to work with Multi-Arch
libefreet1 and efreetd communicate via d-bus, and there is a circular
dependency between libefreet1 and efreetd, which make things difficult
with Multi-Arch (as the helper binary had to be in a separate package).
To avoid that circular dependency issue, libefreet1 is now a
transitional package to libefreet1a and libefreet-bin and the
libefreet1a shlibs file forces a Depends: libefreet-bin. Thanks to
Josselin Mouette for the suggestion!