ecore_evas_convert: make sure we add all needed flags and deps for the build
We got a report where ecore_evas_convert linked against some old libs
installed in the system instead of the recent ones in-tree. While I was not
able to reproduce this the flags and deps for ecore_evas_convert have been
missing in any case and could be the culprit.

Ref T2716